Conditional async session tests
